Update view-analytics.md
I'd like to submit a suggestion. Is it possible to add support - or lack therof - for DoD/GCC-H tenants? 

I opened a Microsoft case back in October (Case# 2310170040009546) and was told this feature wasn't supported for DoD customers, but it doesn't specify here that this feature isn't supported for the DoD.