Update view-analytics.md
Would this please be updated to include the above changes? We opened a Microsoft Case 2310170040009546 back in October 2023 and they stated this wasn't supported or available for users in DoD and GCC-H tenants. We'd like to get official documentation updated with this verbiage to point users to reference regarding this feature.